+++ src/PrettierEditProvider.ts
@@ -22,7 +22,45 @@ import {
 } from './types.d';
 
 const bundledPrettier = require('prettier') as Prettier;
+/**
+ * HOLD style parsers (for stylelint integration)
+ */
+const STYLE_PARSERS: ParserOption[] = ['postcss', 'css', 'less', 'scss'];
+/**
+ * Check if a given file has an associated prettierconfig.
+ * @param filePath file's path
+ */
+async function hasPrettierConfig(filePath: string) {
+    return (await bundledPrettier.resolveConfig(filePath)) !== null;
+}
 
+/**
+ * Define which config should be used.
+ * If a prettierconfig exists, it returns itself.
+ * It merges prettierconfig into vscode's config (editorconfig).
+ * Priority:
+ * - additionalConfig
+ * - prettierConfig
+ * - vscodeConfig
+ * @param hasPrettierConfig a prettierconfig exists
+ * @param additionalConfig config we really want to see in. (range)
+ * @param prettierConfig prettier's file config
+ * @param vscodeConfig our config
+ */
+function mergeConfig(
+    hasPrettierConfig: boolean,
+    additionalConfig: Partial<PrettierConfig>,
+    prettierConfig: Partial<PrettierConfig>,
+    vscodeConfig: Partial<PrettierConfig>
+) {
+    return hasPrettierConfig
+        ? Object.assign(
+              { parser: vscodeConfig.parser }, // always merge our inferred parser in
+              prettierConfig,
+              additionalConfig
+          )
+        : Object.assign(vscodeConfig, prettierConfig, additionalConfig);
+}
 /**
  * Format the given text with user's configuration.
  * @param text Text to format
@@ -32,7 +70,7 @@ const bundledPrettier = require('prettier') as Prettier;
 async function format(
     text: string,
     { fileName, languageId, uri }: TextDocument,
-    customOptions: object
+    customOptions: Partial<PrettierConfig>
 ): Promise<string> {
     const vscodeConfig: PrettierVSCodeConfig = getConfig(uri);
     const localPrettier = requireLocalPkg(fileName, 'prettier') as Prettier;
@@ -41,17 +79,10 @@ async function format(
         return text;
     }
 
-    /*
-    handle trailingComma changes boolean -> string
-    */
-    let trailingComma = vscodeConfig.trailingComma;
-    if (trailingComma === true) {
-        trailingComma = 'es5';
-    } else if (trailingComma === false) {
-        trailingComma = 'none';
-    }
-
-    const dynamicParsers = getParsersFromLanguageId(languageId, localPrettier.version);
+    const dynamicParsers = getParsersFromLanguageId(
+        languageId,
+        localPrettier.version
+    );
     let useBundled = false;
     let parser: ParserOption;
 
@@ -72,24 +103,28 @@ async function format(
         lang.parsers.includes(parser)
     );
 
-    const fileOptions = await bundledPrettier.resolveConfig(fileName);
+    const hasConfig = await hasPrettierConfig(fileName);
 
-    const prettierOptions = Object.assign(
-        {
-            printWidth: vscodeConfig.printWidth,
-            tabWidth: vscodeConfig.tabWidth,
-            singleQuote: vscodeConfig.singleQuote,
-            trailingComma,
-            bracketSpacing: vscodeConfig.bracketSpacing,
-            jsxBracketSameLine: vscodeConfig.jsxBracketSameLine,
-            parser: parser,
-            semi: vscodeConfig.semi,
-            useTabs: vscodeConfig.useTabs,
-            proseWrap: vscodeConfig.proseWrap,
-        } as PrettierConfig,
-        customOptions,
-        fileOptions
-    );
+    if (!hasConfig && vscodeConfig.requireConfig) {
+        return text;
+    }
+    const fileOptions = await bundledPrettier.resolveConfig(fileName, {
+        editorconfig: true,
+    });
+
+    const prettierOptions = mergeConfig(hasConfig, customOptions, fileOptions, {
+        printWidth: vscodeConfig.printWidth,
+        tabWidth: vscodeConfig.tabWidth,
+        singleQuote: vscodeConfig.singleQuote,
+        trailingComma: vscodeConfig.trailingComma,
+        bracketSpacing: vscodeConfig.bracketSpacing,
+        jsxBracketSameLine: vscodeConfig.jsxBracketSameLine,
+        parser: parser,
+        semi: vscodeConfig.semi,
+        useTabs: vscodeConfig.useTabs,
+        proseWrap: vscodeConfig.proseWrap,
+        arrowParens: vscodeConfig.arrowParens,
+    });
 
     if (vscodeConfig.eslintIntegration && doesParserSupportEslint) {
         return safeExecution(
@@ -108,7 +143,7 @@ async function format(
         );
     }
 
-    if (vscodeConfig.stylelintIntegration && parser === 'postcss') {
+    if (vscodeConfig.stylelintIntegration && STYLE_PARSERS.includes(parser)) {
         const prettierStylelint = require('prettier-stylelint') as PrettierStylelint;
         return safeExecution(
             prettierStylelint.format({
@@ -125,9 +160,9 @@ async function format(
         return safeExecution(
             () => {
                 const warningMessage =
-                    `prettier@${localPrettier.version} doesn't support ${
-                        languageId
-                    }. ` +
+                    `prettier@${
+                        localPrettier.version
+                    } doesn't support ${languageId}. ` +
                     `Falling back to bundled prettier@${
                         bundledPrettier.version
                     }.`;
@@ -182,7 +217,10 @@ class PrettierEditProvider
         return this._provideEdits(document, {});
     }
 
-    private _provideEdits(document: TextDocument, options: object) {
+    private _provideEdits(
+        document: TextDocument,
+        options: Partial<PrettierConfig>
+    ) {
         if (!document.isUntitled && this._fileIsIgnored(document.fileName)) {
             return Promise.resolve([]);
         }